Revert "update"
This reverts commit 77823aafbd5a00e5da344c05e2a98e0f00334e7b.

+++ src/main/java/com/greenfarm/dto/CartDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CartDTO {
     private Integer cartID;
-    private Float quantity;
+    private Integer quantity;
      
     Product product;
     User user;
