+++ src/main/java/com/greenfarm/dto/ProductDTO.java
@@ -3,6 +3,7 @@ package com.greenfarm.dto;
 import java.util.List;
 
 import com.greenfarm.entity.Category;
+import com.greenfarm.entity.ProductImage;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -19,7 +20,7 @@ public class ProductDTO {
     private String description;
     private Double price;
     private String image;
-    private Float quantityavailable;
+    private Integer quantityavailable;
 	private Boolean isdeleted = Boolean.FALSE;
     private Category category;
     private List<ProductImageDTO> productimage;
